+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - UTF8 helpers: `str.utf8Len`, `str.utf8Codepoints`, `str.utf8Valid`
 - New integer literal for single chars: `'A' == 65`
 - Compiler will warn you when a local or global is never used or when an expression value is discarded. To silence those warnings you can use the `_ = <expression>` or name the local/global `_`.
+- `std.currentFiber`, `fiber.isMain`
 
 ## Changed